refactor: move verbose to shared
